Rearrangements
One more step towards a less bloated StoneMask.cs
